+++ sand-core/src/lib.rs
@@ -247,6 +247,7 @@ pub use sand_components::{
     AdvancementDisplay,
     AdvancementFrame,
     AdvancementIcon,
+    AdvancementId,
     AdvancementRewards,
     AdvancementTrigger,
     // Custom item
@@ -329,6 +330,7 @@ pub use sand_components::{
     LootFunction,
     LootPool,
     LootTable,
+    LootTableId,
     LootTableType,
     NoiseSettings,
     NumberProvider,
@@ -345,6 +347,7 @@ pub use sand_components::{
     RawComponent,
     RawJson,
     RawSnbt,
+    RecipeId,
     RecipeResult,
     ShapedRecipe,
     ShapelessRecipe,